Use ReactInjection instead of monkey-patching ReactDOMComponent
