+++ src/inject.js
@@ -1,0 +1,4 @@
+import DOMComponent from './DOMComponent';
+import ReactInjection from 'react/lib/ReactInjection';
+
+ReactInjection.NativeComponent.injectGenericComponentClass(DOMComponent);
