+++ src/mixin.js
@@ -1,7 +1,7 @@
 import { PropTypes } from 'react';
 
 import { CLASSMAP_KEY } from './constants';
-import './hook';
+import './inject';
 
 export default function ClassMapMixin(map) {
   return {
